// admin/src/modulos/api/model/medico.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DocumentoIdentidade } from './documentoIdentidade';
import { Especialidade } from './especialidade';
import { ListaEmails } from './listaEmails';
import { ListaEnderecos } from './listaEnderecos';
import { ListaTelefones } from './listaTelefones';
import { MedicoComercial } from './medicoComercial';
import { MedicoCurriculo } from './medicoCurriculo';
import { RedesSociais } from './redesSociais';


export interface Medico {
    /**
     * Código único do médico do sistema
     */
    id?: string;
    /**
     * Código CPF do médico
     */
    cpf: string;
    /**
     * Nome do médico
     */
    nome: string;
    /**
     * Celular principal do médico, será usado para validar ou assinar receitas
     */
    celular: string;
    /**
     * e-mail principal do médico, será usado para validações e informes
     */
    email: string;
    /**
     * Data de nascimento do médico
     */
    nascimento?: string;
    crm: DocumentoIdentidade;
    especialidades: Array<Especialidade>;
    telefones?: ListaTelefones;
    emails?: ListaEmails;
    enderecos?: ListaEnderecos;
    redesSociais?: RedesSociais;
    curriculo?: MedicoCurriculo;
    comercial?: Array<MedicoComercial>;
    /**
     * string base64 contendo o conteúdo do arquivo de certificado digital padrão A1 que será usado para assinar as receitas
     */
    certificadoA1?: string;
    /**
     * string MD5 com a senha do cliente para acesso (aceito apenas de alguns ip/sites)
     */
    senha?: string;
    /**
     * ID único do usuário
     */
    user?: string;
    /**
     * string base64 contendo a foto do médico
     */
    foto?: string;
}