// admin/src/modulos/api/model/redesSociais.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface RedesSociais {
    /**
     * slug do perfil do linkedin
     */
    linkedin?: string;
    /**
     * slug do perfil do Facebook
     */
    facebook?: string;
    /**
     * slug do perfil do Instagram
     */
    instagram?: string;
    /**
     * slug do perfil do Snapchat
     */
    snapchat?: string;
}