// admin/src/modulos/api/model/medicoCurriculo.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Currículo do médico
 */
export interface MedicoCurriculo {
    /**
     * Cadastrar histórico profissional (cursos, certificados, especializações, associações, participações em conselhos, etc).
     */
    descricao?: string;
    /**
     * Lista de títulos
     */
    titulos?: Array<string>;
}