// admin/src/modulos/api/model/medicoComercial.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Endereco } from './endereco';
import { ListaEmails } from './listaEmails';
import { ListaSites } from './listaSites';
import { ListaTelefones } from './listaTelefones';


export interface MedicoComercial {
    /**
     * Nome da clínica, hospital, espaço, etc...
     */
    local?: string;
    /**
     * Titulação do médico no clínica, hospital, espaço, etc...
     */
    cargo?: string;
    endereco?: Endereco;
    emails?: ListaEmails;
    telefones?: ListaTelefones;
    sites?: ListaSites;
}